gaia_object_detection-master/clone_codes/requirements.txt
absl-py==0.7.1
astor==0.6.2
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.268
backcall==0.1.0
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bleach==1.5.0
boto==2.49.0
boto3==1.10.0
botocore==1.13.4
bpemb==0.3.0
cachetools==3.1.1
certifi==2018.4.16
cffi==1.13.2
chardet==3.0.4
cheroot==6.5.5
CherryPy==18.1.2
Click==7.0
cloudpickle==0.5.3
colorama==0.3.9
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.28.3
dartsclone==0.6
dask==0.17.5
decorator==4.3.0
Deprecated==1.2.6
docutils==0.14
easydict==1.9
entrypoints==0.2.3
enum34==1.1.6
feedparser==5.2.1
flair==0.4.4
future==0.17.1
gast==0.2.2
gensim==3.8.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.25.0
h5py==2.8.0
hdf5storage==0.1.15
hickle==3.3.2
html5lib==0.9999999
hyperopt==0.1.2
idna==2.6
importlib-metadata==0.18
ipykernel==5.1.3
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jaraco.functools==2.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.6
jupyterlab-widgets==0.6.15
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
kytea==0.1.4
langdetect==1.0.7
lmdb==0.94
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.1.2
mistune==0.8.3
more-itertools==7.1.0
mpld3==0.3
mysqlclient==1.4.2.post1
natto-py==0.9.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.4.4
notebook==5.5.0
numpy==1.17.3
oauthlib==3.1.0
opencv-python==3.4.1.15
opt-einsum==3.1.0
packaging==19.0
pandocfilters==1.4.2
parso==0.2.1
Pattern==3.6
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.12.0
portend==2.5
prompt-toolkit==2.0.2
protobuf==3.11.1
ptyprocess==0.5.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.2.0
pymongo==3.8.0
pyparsing==2.2.0
pyrsistent==0.15.3
pytest==5.0.1
python-dateutil==2.7.3
python-docx==0.8.10
pytorch-pretrained-bert==0.6.2
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
rdflib==4.2.2
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.2.1
sacremoses==0.0.35
scikit-image==0.14.0
scikit-learn==0.20.2
scipy==1.1.0
segtok==1.5.7
Send2Trash==1.5.0
sentencepiece==0.1.82
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.8.4
sortedcontainers==2.1.0
soupsieve==1.9.2
sqlitedict==1.6.0
SudachiPy==0.4.1
tabulate==0.8.3
tempora==1.14.1
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==2.0.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tiny-tokenizer==3.0.1
toolz==0.9.0
torch==1.3.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.32.2
traitlets==4.3.2
transformers==2.2.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zc.lockfile==1.4
zipp==0.5.2
